% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphaEst.R
\name{alphaEst}
\alias{alphaEst}
\title{Estimation of the overdispersion parameter for Negative Binomial non-negative matrix factorization}
\usage{
alphaEst(data, k, patient_specific = FALSE)
}
\arguments{
\item{data}{Numeric matrix of mutational counts data. Matrix size: no. of mutation types x no. of patients.}

\item{k}{Number of signatures to be used for the non-negative matrix factorization}

\item{patient_specific}{Logical. If TRUE patient-specific overdispersion is used in the Negative Binomial model.}
}
\value{
Overdispersion parameter. Either vector of length one or vector of length no. of patients if patient-specific overdispersion is used.
}
\description{
Empirical Bayes estimation of the overdispersion parameter for Negative Binomial non-negative matrix factorization.
The overdispersion parameter can be either vector of length one or vector of length no. of patients if patient-specific overdispersion is used.
}
