% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA21}
\alias{BRCA21}
\title{Mutational counts for 21 breast cancer patients}
\format{
A data matrix with 21 rows and 96 columns.
}
\source{
\url{https://doi.org/10.1016/j.cell.2012.04.023}
}
\usage{
BRCA21
}
\description{
A dataset containing the mutational counts of 96 different mutation types
for 21 breast cancer patients.
}
\keyword{datasets}
