% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrSignatures.R
\name{corrSignatures}
\alias{corrSignatures}
\title{Correlation of mutational signatures}
\usage{
corrSignatures(H1, H2)
}
\arguments{
\item{H1}{Numeric matrix of mutational signatures.}

\item{H2}{Numeric matrix of mutational signatures.}
}
\value{
Vector of indexes to reorder the second matrix to match the first one.
}
\description{
Correlation of mutational signatures
}
