% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CrossValidation}
\alias{CrossValidation}
\title{Cross-validation for the number of mutational signatures}
\usage{
CrossValidation(
  data,
  k = NULL,
  n_iterations = 100,
  method = "NB",
  cost_f = "GKL",
  size_train = 0.9,
  patient_specific = FALSE,
  tol = 1e-05
)
}
\arguments{
\item{data}{Numeric matrix of mutational counts data. Matrix size: no. of mutation types x no. of patients.}

\item{k}{Number of signatures to be used in the cross-validation.}

\item{n_iterations}{Number of iterations for cross-validation procedure. Default is 100.}

\item{method}{Non-negative matrix factorization model: 'Poisson' or 'NB' can be used. Default is 'NB' corresponding to the Negative Binomial model.}

\item{cost_f}{Cost function to be used for the cross-validation error. Default is GKL (Generalized Kullback-Leibler). Alternatives are: Frobenius norm ("Frobenius") and Itakura-Saito divergence ("IS").}

\item{size_train}{Size of the train set. Default is 0.9, i.e. 90\% of the patients are used for training and 10\% for testing in the cross-validation.}

\item{patient_specific}{Logical. If TRUE patient-specific overdispersion is used in the Negative Binomial model.}

\item{tol}{Threshold for convergence of the EM and MM algorithms for estimating mutational signatures. Default is 1e-5.}
}
\value{
List of length two. 'cost' is a vector of length n_iterations where the cost for each iteration is stored.
cost_k is the median of the values in 'cost' and it is used for comparison to choose the best number of signatures for a given data set.
}
\description{
Cross-validation algorithm for a given number of signatures k using either Poisson or Negative Binomial model.
}
