% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVmodsel.R
\name{CVmodsel}
\alias{CVmodsel}
\title{Cross-validation and model selection pipeline for learning of mutational signatures}
\usage{
CVmodsel(
  data,
  k = 2:3,
  n_iterations = 100,
  cost_f = "GKL",
  size_train = 0.9,
  patient_specific = FALSE,
  tol = 1e-05
)
}
\arguments{
\item{data}{Numeric matrix of mutational counts data. Matrix size: no. of mutation types x no. of patients.}

\item{k}{Vector of integers for the number of signatures values to be tested in cross-validation.}

\item{n_iterations}{Number of iterations for cross-validation procedure.}

\item{cost_f}{Cost function to be used for the cross-validation error. Default is GKL (Generalized Kullback-Leibler). Alternatives are: Frobenius norm ("Frobenius") and Itakura-Saito divergence ("IS").}

\item{size_train}{Size of the train set. Default is 0.9, i.e. 90\% of the patients are used for training and 10\% for testing in the cross-validation.}

\item{patient_specific}{Logical. If TRUE patient-specific overdispersion is used in the Negative Binomial model.}

\item{tol}{Threshold for convergence of the EM and MM algorithms for estimating mutational signatures. Default is 1e-5.}
}
\value{
List of non-negative matrix factorization and residuals results for Poisson model if only the Poisson model is used.
If the Negative Binomial model is used too returns a list of length two with one element per model (Poisson and NegBin). Each of them is a list of two elements for the results of the matrix factorization and the residuals.
}
\description{
Analysis pipeline with model selection and cross-validation for extracting mutational signatures from
mutational counts data using non-negative matrix factorization. A Poisson model is tested first and, if
overdispersion is observed, then a Negative Binomial model is tested.
}
