% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFNB.R
\name{NMFNB}
\alias{NMFNB}
\title{Negative Binomial non-negative matrix factorization}
\usage{
NMFNB(M, N = NULL, alpha, tol = 0.001, seed = sample(1:1000, 1))
}
\arguments{
\item{M}{Numeric matrix of mutational counts data. Matrix size: no. of mutation types x no. of patients.}

\item{N}{Integer. Number of signatures to be used for matrix factorization.}

\item{alpha}{Overdispersion parameter. Either vector of length one or vector of length no. of patients if patient-specific overdispersion is used.}

\item{tol}{Threshold for convergence of the MM algorithm for estimating mutational signatures. Default is 1e-5.}

\item{seed}{Vector of seeds for initializing the values of the signatures and exposures matrices. If default only one initialization is performed.}
}
\value{
List of non-negative matrix factorization results. P is the signature matrix and E is the exposures matrix
}
\description{
MM algorithm for negative binomial non-negative matrix factorization
}
