% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFPoisEM.R
\name{NMFPoisEM}
\alias{NMFPoisEM}
\title{Poisson non-negative matrix factorization}
\usage{
NMFPoisEM(M, N = NULL, tol = 1e-05, seed = sample(1:1000, 1))
}
\arguments{
\item{M}{Numeric matrix of mutational counts data. Matrix size: no. of mutation types x no. of patients.}

\item{N}{Integer. Number of signatures to be used for matrix factorization.}

\item{tol}{Threshold for convergence of the EM algorithm for estimating mutational signatures. Default is 1e-5.}

\item{seed}{Vector of seeds for initializing the values of the signatures and exposures matrices. If default only one initialization is performed.}
}
\value{
List of non-negative matrix factorization results. P is the signature matrix and E is the exposures matrix
}
\description{
EM algorithm for poisson non-negative matrix factorization
}
