% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigMoS.R
\name{sigmos}
\alias{sigmos}
\title{Model selection algorithm for the number of mutational signatures}
\usage{
sigmos(
  data,
  k = NULL,
  n_iterations = 100,
  method = "NB",
  cost_f = "GKL",
  size_train = 0.9,
  patient_specific = FALSE,
  tol = 1e-05,
  beta = 2
)
}
\arguments{
\item{data}{Numeric matrix of mutational counts data. Matrix size: no. of patients x no. of mutation types.}

\item{k}{Number of signatures to be used in the cross-validation.}

\item{n_iterations}{Number of iterations for cross-validation procedure. Default is 100.}

\item{method}{Non-negative matrix factorization model: 'Poisson' or 'NB' can be used. Default is 'NB' corresponding to the Negative Binomial model.}

\item{cost_f}{Cost function to be used for the cross-validation error. Default is GKL (Generalized Kullback-Leibler). Alternatives are: Frobenius norm ("Frobenius"), Itakura-Saito divergence ("IS"), Beta divergence ("Beta").}

\item{size_train}{Size of the train set. Default is 0.9, i.e. 90\% of the patients are used for training and 10\% for testing in the cross-validation.}

\item{patient_specific}{Logical. If TRUE patient-specific overdispersion is used in the Negative Binomial model.}

\item{beta}{Coefficient for Beta divergence if Beta divergence is used as cost function. Default is 2.}
}
\value{
List of length two including the costs.
\itemize{
\item \strong{cost} is a vector of length n_iterations where the cost for each iteration is stored.
\item \strong{cost_k} is the median of the values in 'cost' and it is used for comparison to choose the best number of signatures for a given data set.
\item \strong{FullDataCost} is the likelihood value for the fit of the full data
\item \strong{Signatures} The signature matrix for the full data
\item \strong{Exposures} The exposure matrix for the full data
}
}
\description{
Model selection algorithm for a given number of signatures k using either Poisson or Negative Binomial model.
}
\examples{
# Use SigMoS with the Negative Binomial distribution:
res <- sigmos(BRCA21,k=3,patient_specific = TRUE)
str(res$Signatures)
str(res$Exposures)
# Use SigMoS with the Poisson distribution:
res <- sigmos(BRCA21,k=3,method="Poisson")
str(res$Signatures)
str(res$Exposures)
#### Not Run
## Find estimated number of signatures for the given example using the Negative
## Binomial distribution. Evaluate no. of signatures between 2 and 7:
# res_cv <- list()
# CVcost <- rep(0,6)
# for (i in 2:7){
#  res_cv[[i]] <- sigmos(BRCA21,k=i,method="NB")
#  CVcost[i-1] = res_cv[[i]]$cost_k
# }
# which.min(CVcost)+1 #estimated number of signatures

}
