% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SimulatedDataNBn}
\alias{SimulatedDataNBn}
\title{Simulated data with patient specific Negative Binomial noise}
\format{
A list of 100 elements where each element is a matrix with 100
rows and 96 columns.
}
\source{
Details on the simulation setup are available at \url{ourpaper}
}
\usage{
SimulatedDataNBn
}
\description{
100 simulated mutational counts data sets. Each data set contains 100 patients, 96 different mutation types.
}
\details{
The true number of underlying signatures for each data set equals to 5.

We added Negative Binomial noise to the simulated mutational counts
data with alpha values inspired from the ones of the patients in
\url{https://doi.org/10.1016/j.cell.2012.04.023} (i.e. patient
specific overdispersion).
}
\keyword{datasets}
